refactor: simplify parsing internals
