+++ src/utils.ts
@@ -55,3 +55,9 @@ export function factorize(integer: number): Record<number, number> {
 
 	return factors;
 }
+
+export const isObject = <T = Record<string, unknown>>(input: unknown): input is T => {
+	return typeof input === "object"
+		&& !Array.isArray(input)
+		&& input !== null;
+};
